API/requirements/common.txt
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
PyMySQL==0.10.1
cryptography==3.2.1
pika==1.1.0